1. AuthImg 엔티티 @Id Long 타입으로 변경 2. GoalApplication에 @EnableJpaRepositories 어노테이션 추가

+++ src/main/java/com/allrounders/goalkeeper/domain/AuthImg.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 public class AuthImg {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer authImgId;
+    private Long authImgId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "memberId")   // foreign key (member_id) references Member (member_id)
